Move ElementTreeNode to its own file
